requirements/local.txt
-r base.txt
django-debug-toolbar
pytest
pytest-django
factory-boy
faker
